requirements.txt
numpy 
pandas 
matplotlib 
seaborn 
streamlit 
google-generativeai
scikit-learn 
scipy